Reduce size of stored snapshot images
We have some very large image snapshots in the database (the 10 largest
images are between 18MB and 350MB).

Storing the images as png with a quality of 1 appears to give the
highest quality for the lowest file size.

This will only affect new snapshots. I'll resize the existing ones
separately.

There's more information in
https://trello.com/c/Nbk1otVM/212-very-large-image-screenshots.