Ensure Rails server detects code changes when running in dev-vm
I was having to manually restart the server when making changes in the
dev-vm. The fix appears to be use the `FileUpdateChecker` instead of the
`EventedFileUpdateChecker`.

I only want to use the `FileUpdateChecker` when running in the VM so I'm
writing and then checking the `RUNNING_IN_VAGRANT` environment variable,
as suggested by this answer on Reddit[1].

[1]: https://www.reddit.com/r/vagrant/comments/74tzt9/how_to_set_environment_variable_on_guest_machine/