Avoid building snapshot when the URL is broken
The `Statement#fetch_snapshot` method would try to build a snapshot when
the response from the request was bad (e.g. a 404) if an admin had
previously marked the statement's URL as not being bad (i.e. set
`marked_not_broken_url` to true). My guess is that this wasn't the
intention when the functionality was added in
44a3ea53245c985bb6a1277a711203f0472c1c82: it doesn't make sense to store
a snapshot because the content data is either empty (leading to a number
of exceptions being reported in Rollbar) or, at best, present but not
useful.

I've had to update the 'When Patricia marks the URL for "Cucumber Ltd"
as not broken' step so that it returns `broken_url: false` for requests
made using `StatementUrl` and for requests made using `ScreenGrab`.