Enable pg_trgm extension in database
Enable the `pg_trgm` extension[1] which allows us to calculate the
similarity between two strings based on trigram matching. We can use
this to make the search engine a bit more "fuzzy".

This extension appears to be available in Heroku[2] and RDS[3] so we
should be safe to enable it now and when we migrate to AWS.

[1] https://www.postgresql.org/docs/9.1/static/pgtrgm.html
[2]
https://devcenter.heroku.com/articles/heroku-postgres-extensions-postgis-full-text-search
[3] https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/CHAP_PostgreSQL.html