Use the pg_trgm similarity matcher for company name searches
See: https://trello.com/c/gWCGdmc5/172-variations-in-searches

This should improve the case when, for example 'Cucumber Ltd' appears
in the database but the user searches for 'Cucumber Limited'.

It's hard in general to know whether this improves the ease of use for
all users and search terms, but as it seems to improve the specific
problem mentioned in the trello card I think it's worth deploying and
seeing how it goes.