Fix bug when displaying 'X' in industry and country filter
https://trello.com/c/vVIIKyqV/201-styles-on-explore-page-filters

I wasn't able to reproduce this bug locally but we see it in
production.

According to https://github.com/tsechingho/chosen-rails/issues/97 this
is a known issue in `chosen-rails` but as of Feb 2018 there hasn't
been any sign that the fix proposed in the issue will be
released. Hopefully the fix introduced in this commit can be reverted
at some point.