Use postgres full text search instead of trigrams
https://trello.com/c/gWCGdmc5/172-search-function

This commit switches the company name search to use postgres full text
search instead of the pg_trgm search introduced in 9bac1a0. The
problem with the latter is that although it handles misspellings quite
well it doesn't cope with very short search terms.

The current implementation is slow as the name columns are converted to
vectors at query time. I'll look into improving things in subsequent
commits.