Substitute common abbreviation of 'limited' in company search
https://trello.com/c/gWCGdmc5/172-search-function

Currently when users search for, for example "nike ltd" the search
results are empty. This is because the company name is "nike (uk)
limited".

This commit adds a search alias from 'ltd' to 'limited' and
vice-versa.

To do that we have added a new table `search_aliases` following the
pattern given in the Postgres documentation[1]. Unfortunately
activerecord does not support the `tsquery` type so we cannot use a
standard rails migration and have to resort to using raw SQL.

At the moment we have no need to create a corresponding Rails model
for the search_aliases table, but to avoid future confusion I have
added an auto-incrementing id column to `search_aliases` so that it is
easier to use a model if required.

[1] https://www.postgresql.org/docs/8.3/static/textsearch-features.html#TEXTSEARCH-QUERY-REWRITING