Add BHRRC URL to companies
This is admin-only functionality at the moment. It allows admins to
enter/view the URL of the company on the BHRRC website.

We'll want to show this somewhere in the public-facing site but I need
some more information before making that change.

I feel slightly uncomfortable about not adding any tests in this commit.
I couldn't find any Cucumber features to extend and there aren't
currently any controller specs for the companies controller.