Refactor calculation of statements over time
I found the calculation of the cumulative number of statements
using a combination of ruby and activerecord queries confusing. I
personally find it easier to read as a single SQL expression.

To do the refactor I've ensured that every statement in the setup is
assigned at least one Legislation and clarified that the expected
total number of statements is therefore the unique number of
statements. This change will allow me to more easily group by
legislation in a subsequent commit.