Add number of statements published under each act to stats hash
This commit adds `uk_act` and `us_act` fields to the hash returned
from StatementStats#total_statements_over_time. The number is intended
to reflect the count of statements published under each act per
month.

A statement can count twice if it's published under two different
acts.

At the moment the query is hard-coded to the exact names of the two
acts that are in the production `legislations` table. This means
we return no stats unless the two production legislations exist.