Indicate whether a statement should have a screenshot
This should make it easier to understand why a preview might be missing
for a particular statement. We can only display previews for HTML
statements that we were able to generate a screenshot for, or that are
available via HTTPS. There wasn't an easy way to know whether we'd
generated a screenshot for a statement so it wasn't particularly easy to
understand why a statement might be missing.

Related Trello card: https://trello.com/c/gV3GW6j4