Avoid exception if statement submitted with blank URL
The exception was `undefined method `each' for nil:NilClass` when trying
to iterate over `@statements` in the _company_tabs.html.erb partial.

I've taken the opportunity to rename `@statement` to `@new_statement`
(to match the name in the `new` action) so that the "Submit new
statement" tab is highlighted as expected.