Show statements & companies per legislation on the homepage
To make it clearer that the website deals with UK, California, and soon
Australian statements.

The Australian flag image is from
https://github.com/lipis/flag-icon-css. I'm not sure where the UK/US
flag images came from but they look similar to those available on this
site so I'm guessing this may have been the source.

Note that this change makes the numbers look slightly higher because a
statement filed under both UK and California legislation will now be
counted twice where it was previously only counted once in the figures
we're replacing.

I've tried various things to get the stats and "Coming soon" link to
line up vertically but haven't been able to because of the varying
lengths of legislation headings.

I've had to update the download_steps.rb so that the statements are
downloaded from the explore page now that the equivalent link has been
removed from the homepage.