Extract California legislation name
To DRY up use of "California Transparency in Supply Chains Act" in the
app.

I'm not particularly comfortable with us relying on the presence of a
string in the database; but that was already the case. This change
simply means that we only have to update the string in a single place if
it changes in the database.