Cache pages and nav-links to reduce memory overhead
In several places in the app, we were selecting the
whole page table into memory, marshalling it to objects
in order to make navigation. In particular this was
happening on every page in the header and footer, and
on the home page when deciding whether to show the banner.

The page table only has 4 or so rows in it, but each row
contains the HTML for a page as text which can run to many
kilobytes. There's a suspicion that page loading might be
behind the slow start-up/release times we see on AWS and
the memory errors that we see on Heroku.

This commit adds some caching methods in the application_helper
to provide navigation and all the pages. In order to allow the
cache to be long-lived and expire when new pages are created
or old ones changed, this commit also adds timestamps to the
pages object and a class method to Page that gives you the
date of the most recent change to the pages table.