Redirect to snapshots stored on S3
Instead of downloading them and serving them from the Rails app. I hope
this will decrease the amount of RAM the app is using on Heroku.

Calling `url_for` with `snapshot.screenshot_or_original` generates an
app-specific URL for the asset which, when requested, returns a
temporary, signed URL for the asset on S3[1].

Although the time-limited S3 URL can be discovered (by right-clicking on
a snapshot image/PDF and opening it in a new tab) I don't think this is
going to be a problem.

I've tested this in Chrome and the external behaviour (i.e. as far as a
user is concerned) remains the same.

[1]: http://guides.rubyonrails.org/active_storage_overview.html#linking-to-files