Include other companies covered by a statement in search results
This commit adds the text in `statements.also_covers_companies` to the
word vector used when building a search.

In practice this means that searches for a company that is covered by
an existing statement will return that statement. For example, before
this commit the following query returned no results:

    https://www.modernslaveryregistry.org/explore?company_name=acorn+recruitment

Whereas after this commit this query returns results for the Acorn
Group[1] because their statement also covers "Acorn Recruitment Ltd".

[1] https://www.modernslaveryregistry.org/companies/6899-acorn-group/statements/27735