Add extension to configure sidekiq on the server
We used a Worker dyno to run the Sidekiq process on Heroku. It's not
quite so simple in Elastic Beanstalk: while it does have support for
Worker environments they're designed to use SQS and to not necessarily
have access to the code in the web app. I've decided to run the Sidekiq
workers on the EC2 instances that are hosting the web application for
now.

This script is mostly copied from
https://openclassrooms.com/en/courses/4567631-deploy-rails-applications/4794886-configure-elastic-beanstalk-to-run-sidekiq.