Don't report ActionController::UnknownFormat to Rollbar
We see a number of these exceptions in Rollbar in response to requests
for formats we don't support (e.g. text/plain). The user receives a 406
Not Acceptable response but the error is still logged unnecessarily. I
think it's OK to ignore these exceptions because they're also ignored by
default by Honey Badger[1].

[1]: https://github.com/honeybadger-io/honeybadger-ruby/blob/916df6b731a70532a27d4e94c404361552e6861e/lib/honeybadger/config/defaults.rb#L12