Remove `data-content` attribute which breaks a feature spec
In 1f188e9 we introduced a new column and gave it the `data-content`
attribute `stats_counters`. This broke the
`features/public_stats.feature` feature as it relies on the
`data-content` to target a selector.

This attribute doesn't seem required for the layout and I suspect it
might have been introduced in error. I'm removing it to fix the build.