Add RSS feed for recently published statements
https://trello.com/c/7MDwEkdN/239-rss-feed-of-latest-statements

This commit adds an RSS feed, at `/statements.rss` containing the most
recently published 20 statements.

In the RSS feed I've included the same metadata we see on a statement
page (legislation, date seen and verifying person).

Note that we don't currently track when a statement is published so
I've had to order by creation date instead.

The view spec uses the RSS parser in the standard library to ensure
that the RSS feed is valid and parsable, it felt like a bit too much
work to fully test all of the content.