Add methods to count also covered companies in StatementStats
These methods count the unique count of `also_covered_companies`
across published statements assuming that the field contains a list of
comma separated company names.

This is not the most robust way of counting these companies. Ideally
we'd model the also_covered_companies as an association between the
statement and multiple Company records. We could then count the unique
number of associated Companies instead. Relying on parsing a comma
separated list may result in counting entries such as "N/A" or other
dodgy data.

It would be quite a big job to model the association correctly and I
think this is close enough to achieve the objective of more accurately
reflecting the true number of companies covered by statements.

We can also tidy up the data to remove some of the issues.