Remove Statement#additional_companies_covered
I initially added this field in 49fe298927 with the intention of using
it to count the number of "also_covers_companies" in stats. In
aac1497f we realised that using this count wouldn't correctly unique
the companies. As we are no longer using it (and have an alternative
approach in a8dc916310) this field can be removed.