Calculate compliance stats for published statements only
We noticed that the compliance stats were being calculated for all
latest statements rather than the latest published ones. We spoke to
Alex and he confirmed that this was probably an oversight.

Switching to use the latest_published scope doesn't alter the
calculated values based on yesterday's production data and will allow us
to remove the `latest` column on the statement table.