Change ComplianceStats to not rely on Statement#latest_published
We'd like to remove Statement#latest_published to make it easier to
associate a statement with multiple companies. Removing
the `Statement#latest_published` makes it easier to make that change.

The `latest_published` field on the `statements` table appears to have
been introduced to speed up the calculation of the compliance
stats. We've replaced the use of the field in ComplianceStats with a
SQL query.

We noticed a couple of problems with the previous implementation of
this calculation:
 - Previously it was possible for Statement#latest_published to be set
 to true in the database even if the statement had since been
 unpublished.
 - If the latest published statement was for a non-UK legislation it
 would have been included in the compliance stats by mistake. We've
 added a feature to demonstrate this problem and show that the new
 query fixes it.

We have noticed that the new query is around 2x slower to execute than
the old one, which adds around 20% to the render time of the
homepage in development. We're not too concerned about this
degradation at this stage as it's still acceptable and we plan to
revisit the calculation when we add filtering by industry.