Add _very_ simple healthcheck page
So that we can avoid the ELBs hitting the homepage in order to check the
health of the app. A 200 response from this page should at least
indicate that the app can talk to the database, which is probably about
all the homepage offers; albeit the homepage does it by performing quite
an expensive query to generate compliance stats.

We're seeing some instance health degradation warnings in the Elastic
Beanstalk events and wonder whether it's in part to do with the ELB
making requests to the application homepage. We'll deploy this and see
whether things improve.