Upgrade to bulma-rails 0.5.1
So that I can add a multiple-select dropdown and have it styled in
keeping with the rest of the admin site. Styling multi-select dropdowns
appears to have been added to bulma in 0.4.4 but bulma-rails jumps from
version 0.4.3 to 0.5.1[1].

Additional cosmetic differences:

* Slightly larger padding/margin across all pages which reduces the area
of the content by a small number of pixels.

* Tags[2] are now rectangular with rounded corners instead of eliptical.
It's possible to retain the old behaviour by adding the `is-rounded`
class but I think the new ones look slightly better.

* Tables[3] need the `is-fullwidth` class to retain the style they had in
bulma 0.4.3.

* Table headings are now bold instead of a lighter grey. I think this is
an improvement so have left it.

See the Bulma 0.5.1 docs[4] for more info.

[1]: https://rubygems.org/gems/bulma-rails/versions
[2]: https://versions.bulma.io/0.5.1/documentation/elements/tag/
[3]: https://versions.bulma.io/0.5.1/documentation/elements/table/
[4]: https://versions.bulma.io/0.5.1/documentation/overview/start/