Destroy orphaned statements when destroying companies
To avoid the problem of leaving orphaned statements in the system when
companies are deleted. An orphaned statement prevented access to /admin
earlier today as our code assumes that a statement will always have a
company.

Note that we don't remove the statement if it's still associated with
another company.