Use named route helper methods in link_to
Instead of a combination of helper methods and arrays.

Using the array form makes sense in `form_for` when we want it to work
for new and saved objects but I don't think this applies when calling
`link_to`. Explicitly using the helper methods will also make my life
easier when I make the statement routes shallow in future.

Note the conditional added to the cancel button in
companies/_form.html.erb template. This replicates the magic that was
there previously where if @company wasn't persisted it would redirect to
the index page, and if it was it would redirect to the company page.