Use named route helper methods in button_to
Instead of a combination of helper methods and arrays.

Using the array form makes sense in `form_for` when we want it to work
for new and saved objects but I don't think this applies when calling
`button_to`. Explicitly using the helper methods will also make my
life easier when I make the statement routes shallow in future.