Use named route helper methods instead of array
Using the array form makes sense in `form_for` when we want it to work
for new and saved objects but I don't think this applies when generating
the path for this cancel link. Explicitly using the helper methods will
also make my life easier when I make the statement routes shallow in
future.