Move statement stats graph to stats page
Before this commit the line graph showing statements added by time was
displayed to admins only in the compliance stats carousel on the
homepage. This commit moves the graph to the admin-only "stats" page
and removes it from the homepage. The stats page is a more natural
home for this graph.