Improve performance of ComplianceStats.compile
Using separate SQL COUNT queries rather than a `find_each` loop
reduces the time taken for this query from around 1.1s on my laptop to
0.6s.

This reverses most of the damage done in 1e978d91b80. We're somewhat
hopeful that this change will mean that less application instances
need to be run in AWS, but we're not sure of that yet. This change
seems valuable in either case.