Fix duplicate "Oil, Gas & Consumable Fuels" industry
In 360160a835ea2 we introduced a default set of industries that appear
to be taken from the Global Industry Classification Standard[1].

We have two industries with the name "Oil, Gas & Consumable Fuels"
which appears to be a mistake.

There is no admin interface to modify the industries themselves so I
am fairly confident that we can rely on the IDs of the industries
records to identify them, and have take those ids from a production db
dump made shortly before I wrote this migration.

As the admin selects industries based on the name, I'm confident that
all of the companies that have `industry_id` 2 or 3 were meant to be
classified as "Oil, Gas & Consumable Fuels` and this migration
reassigns the industry_id 2 to all of them (this one has the correct
industry_code 101020 in the DB). I then delete the incorrect industry
record with id 3.

[1] https://en.wikipedia.org/wiki/Global_Industry_Classification_Standard