Improve performance of ComplianceStats#approved_by_board_count
We think the use of `in` with a large array (in the region of 6000
statement IDs) might be causing the problems we're seeing with the site
on AWS.

This change certainly seems to perform better based on some manual
testing in the Rails console.

I've introduced quite a lot of duplication in this commit but hope to
DRY it up in future.