Use legislation ids in explore page query params
This uses the more standard `legislations[]` format of query
parameters to pass to rails and avoids the string manipulations we had
to do when using legislation names.

It also allows the filter to be simplified so that it has the same
form as the other filters.

There's a slightly confusing behaviour here when both legislations
checkboxes are unticked - in this case we don't filter the results at
all and return all results (in a similar way to have the country
filter works when no county is selected). But it might make more sense
to switch the UI to use a 3-way radio button to avoid this confusing
fourth state.