Replace explore page search with CompanySearch
Now we have created Company records for companies listed in the
Statement `also_covers_companies` field we can improve the
explore/search page by searching these company records directly rather
than searching via the statement association.

This allows us to list companies that were previously "also covered"
as entries in the search results and link directly to their own
company page.

This commit replaces the Statement.search method with a new
CompanySearch class that filters by the passed in industry and
country. A change in behaviour here is that we want to display the
company even when it has no statements under the selected
legislation (if selected) to allow us to encourage the visitor to
submit missing statements.

This class has the same public interface as StatementSearch but
currently the `industry_stats` and `stats` methods return dummy
values.

At the moment the "download statements" feature remains unaffected,
we're still using the StatementSearch to provide the results.