Fix intermittently failing examples in compliance_stats_spec
I introduced this mistake when I added the test. I was creating two
published statements for company 1, and therefore industry 1. Because
`last_year_covered` wasn't set, and `date_seen` would've been the same
for both statements, it wasn't possible to be confident that the
industry1_statement was being used when generating the compliance stats.