Remove Company#statements_from_other_companies habtm
We want to add callbacks that are invoked when records are added or
removed from the companies_statements join table. It's easier to do this
if we only define the habtm on one side of the relationship. We only
need to be able to associate a statement with other companies (i.e.
`Statement#additional_companies_covered`) as that's what the UI allows.