Remove unpublished statements from explore
Previously admins were able to see unpublished statements in results
on the /explore page, and in the associated statement download.

This feature makes the refactor we're planning of the explore page to
list companies rather than statements much more complicated. Having
spoken to Alex we've confirmed that the view of unpublished statements
in this interface isn't often used, since admins are able to view all
companies and statements in the admin interface.

I've also removed the unpublished statements in the CSV download for
now so that it is consistent with the same presentation of the data in
the explore page. We may consider adding the unpublished statements
back for admins after the refactor.