Split text extraction and reindexing
It was taking too long and would kill the index process if there was an issue with the text extraction.

So we've split this into two steps an extraction job which will run after the FetchStatementSnapshotJob

We also store the text into the db, and scope processed models, so we do not need to check models that we have already checked.

We do not want to retry the extraction as if there is an issue we need to investigate rather than retry.

At the moment we are using the default queue, might need to split the snapshot and extraction into their own queues.

Also switched to using the original file rather than the screenshot.